Move parseValue to yqlib/value_parser.go
