+++ yq_test.go
@@ -9,24 +9,6 @@ import (
 	"github.com/mikefarah/yq/test"
 )
 
-var parseValueTests = []struct {
-	argument        string
-	expectedResult  interface{}
-	testDescription string
-}{
-	{"true", true, "boolean"},
-	{"\"true\"", "true", "boolean as string"},
-	{"3.4", 3.4, "number"},
-	{"\"3.4\"", "3.4", "number as string"},
-	{"", "", "empty string"},
-}
-
-func TestParseValue(t *testing.T) {
-	for _, tt := range parseValueTests {
-		test.AssertResultWithContext(t, tt.expectedResult, parseValue(tt.argument), tt.testDescription)
-	}
-}
-
 func TestMultilineString(t *testing.T) {
 	testString := `
 	abcd
